[클라/수정/공격] 캐릭터 추적 후 공격을 위한 상태 클래스 구현

+++ HifeSurvival/Assets/Scripts/Realtime/Entity.cs
@@ -30,11 +30,11 @@ public class Stat
     public int hp;
     public int exp;
     public int str;
-    public int attackRange;
+    public float attackRange;
     public int detectRange;
     public float speed;
 
-    public Stat(int inMaxHP, int inMaxEXP, int inMaxSTR, int inDetectRange,  int inAttackRange, float inSpeed)
+    public Stat(int inMaxHP, int inMaxEXP, int inMaxSTR, int inDetectRange,  float inAttackRange, float inSpeed)
     {
         hp = maxHP = inMaxHP;
         exp = maxEXP = inMaxEXP;
